"Two writes" is not sufficient for a singleton -- dereference is
The two-write create/destroy signature produced two false positives as soon as
I tried to resolve the leftovers: BIGHANA_MGR's 0x5B30 is a plain int counter
(the "create" is a decrement) and OBJ_WENDY's 0xD6EC is an int state value
(loaded, compared against 1, assigned 2). Neither points at anything, and
hunting for a class behind either would have dead-ended exactly the way
lbl_2_bss_11B70 did -- from the opposite direction, and just as confidently.

The real test is whether the value LOADED from the label is then used as a base
register. A pointer's loaded value appears as the rA of a following load/store;
a counter's is only compared or arithmetic'd. bss_classify.py now requires that
before reporting a singleton. All seven real singletons still classify; the two
counters no longer do.

The shape of the mistake matters more than the fix: a signature that matched
every example I had was promoted to a rule without being tested against cases
it should reject. Finding out depended on going back for the two rows I had
left marked unresolved -- an unresolved row invites a check, a confidently
wrong row does not.

+++ wip/wm_units/agent_kinopio/d_a_wm_kinopio.cpp
@@ -278,9 +278,15 @@ void daWmKinopio_c::stepCutscene70() {
     switch (m_1a8) {
     case 0:
         if (mPos.x > m_19c.x - 100.0f) {
+            // @unofficial the divisor is a genuine `int 15`, not a float
+            // literal -- confirmed by the target's runtime int-to-float
+            // bit-trick conversion (li r3,0xf; lis r0,0x4330; xoris;
+            // stw/stw; lfd; fsubs) appearing TWICE, which the compiler
+            // would never emit for a compile-time float constant (that
+            // pools as a plain `lfs` immediate instead).
             float dist = (m_19c.x - mPos.x) * 1.0f;
-            float speed = dist / 15.0f;
-            m_194 = speed / 15.0f;
+            float speed = dist / (int) 15;
+            m_194 = speed / (int) 15;
             mpMdlMng->mpMdl->setAnm(2, 3.0f, 5.0f, 0.0f);
             m3d::mdl_c *mdl = mpMdlMng->mpMdl->getBodyMdl();
             m_1b0 = fn_80103520(dWmEffectManager_c::m_pInstance, 2, mdl, "kinopio_all_root", 0, 0);
